Implement StyleSheetList#length
The implementation uses a layout query to get the number of author stylesheets from stylist. That's probably not ideal, but right now it seems to be the only option. In a final version, we probably want to have an association between a document's `HTMLLinkElement`s and `HTMLStyleElement`s and their underlying stylesheets. Then we could iterate over a document's nodes and filter out the ones representing stylesheets. Right now that doesn't seem to exist, though.