Implement basic support for reflecting CSS stylesheets to JS
Mainly, this encompasses storing `dom::CSSStyleSheet` instead of `style::Stylesheet` instances on the document and returning them via the `Item` method on `StyleSheetList`. Ideally, the `CSSStyleSheet` instances would be created lazily upon first use, but that turned out to be quite annoying to implement, so I left it as a follow-up task.

One wrinkle is that we create an internal `style::Stylesheet` for `<meta viewport>` nodes. Right now, this stylesheet is reflected like the ones for `<style>` and `<link rel="stylesheet">` nodes. It should be hidden from content, though. I'm not yet sure whether to implement this using two separate lists on the document - one for `style::Stylesheet`s, one for `dom::CSSStyleSheet`s - or with a custom iterator over the list that omits `<meta viewport>` nodes. The former would make lazy instantiation of the JS reflection simpler, so probably that.